Fix exception restoring with V1 lock files.
When working with V1 lock files, we set ProjectFileToolGroups to null. LockFile.Equals expects it to be non-null, so we get an NRE restoring a project with a V1 lock file.

This change adds an extension method OrderedEquals which can replace the following commonly used pattern:

    this.enumerable.OrderBy(x => f(x), comparer)
        .SequenceEqual(other.enumerable.OrderBy(x => f(x), comparer))

With the following:

    this.enumerable.OrderedEquals(other.enumerable, x => f(x), comparer)

Since this has benefits everywhere we use this pattern (handles null refs, easier to read, reduced code duplication), I applied the change everywhere.

This resolves https://github.com/NuGet/Home/issues/2452